% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pconsumption}
\alias{pconsumption}
\title{Phosphorus consumption in agriculture by region}
\format{
A data frame with 11 rows and 10 variables
\itemize{
  \item Africa: p fertilizer consumption in countries
AGO, BDI, BEN, BFA, CAF, CIV, CMR, COD, COG, DJI, DZA, EGY, ERI, ETH, GAB, GHA, GIN, GMB, GNQ, KEN, LBR, LBY, MAR, MDG, MLI, MOZ, MRT, MUS, MWI, NAM, NER, NGA, RWA, SDN, SEN, SLE, SOM, SWZ, TCD, TGO, TUN, TZA, UGA, ZAF, ZMB, ZWE
  \item East Asia: p fertilizer consumption in countries
CHN, BRN, IDN, JPN, KHM, KOR, LAO, MMR, MNG, MYS, PHL, PRK, SGP, THA, TLS, VNM
  \item Eastern Europe and Central Asia: p fertilizer consumption in countries
ARM, AZE, BLR, EST, GEO, KAZ, KGZ, LTU, LVA, MDA, RUS, TJK, TKM, UKR, UZB
  \item Latin America: p fertilizer consumption in countries
ARG, ATG, BHS, BLZ, BMU, BOL, BRA, BRB, CHL, COL, CRI, CUB, CYM, DMA, DOM, ECU, GRD, GTM, GUY, HND, HTI, JAM, KNA, LCA, MEX, NIC, PAN, PER, PRY, SLV, SUR, TCA, TTO, URY, VCT, VEN
  \item North America: p fertilizer consumption in countries
CAN, USA
  \item Oceania: p fertilizer consumption in countries
AUS, FJI, NCL, NRU, NZL, PLW, PNG, PYF, SLB
  \item South Asia: p fertilizer consumption in countries
BGD, IND, LKA, NPL, PAK
  \item Western and Central Europe: p fertilizer consumption in countries
ALB, AUT, BEL, BGR, BIH, CHE, CZE, DEU, DNK, ESP, FIN, FRA, GBR, GRC, HRV, HUN, IRL, ISL, ITA, LUX, MKD, MNE, NLD, NOR, POL, PRT, ROU, SRB, SVK, SVN, SWE
  \item Western Asia: p fertilizer consumption in countries
AFG, ARE, BHR, CYP, IRN, IRQ, ISR, JOR, KWT, LBN, OMN, PSE, QAT, SAU, SYR, TUR, YEM
}
}
\usage{
pconsumption
}
\description{
This compilation provides records on the regional phosphorus (P) consumption in agriculture
derived from the IFA Consumption database.
}
\details{
Grand Total P2O5 values in the IFA Consumption database. 
Plant nutrition uses only (applications to crops, pastures, forests, fish ponds, turf, ornamentals). 
The consumption estimates provided in the Consumption Database relate, to the extent possible, to real consumption.
P consumption is aggregated by IFA region.

Metric tonnes of P2O5 per annum are used for volumes throughout this compilation.

Countries are reported by ISO 3166 country code.
}
\references{
\itemize{
	 \item IFASTAT Consumption database (2020). Downloaded from https://www.ifastat.org/databases/plant-nutrition on Jun 6th 2020. 	
}
}
\keyword{datasets}
